Refactor: test CollectionUtils -> general collectionUtils.  (#3535)
* (rebase): remove ec2 changes from this branch

* remove TODO comment


* delete final remnants of ec2

* delete old file

---------
