+++ src/test/apprunner/explorer/apprunnerServiceNode.test.ts
@@ -11,7 +11,7 @@ import { AppRunnerNode } from '../../../apprunner/explorer/apprunnerNode'
 import { AppRunnerServiceNode } from '../../../apprunner/explorer/apprunnerServiceNode'
 import { DefaultAppRunnerClient } from '../../../shared/clients/apprunnerClient'
 import { DefaultCloudWatchLogsClient } from '../../../shared/clients/cloudWatchLogsClient'
-import { asyncGenerator } from '../../utilities/collectionUtils'
+import { asyncGenerator } from '../../../shared/utilities/collectionUtils'
 import { AWSTreeNodeBase } from '../../../shared/treeview/nodes/awsTreeNodeBase'
 import { stub } from '../../utilities/stubber'
 import { getLabel } from '../../../shared/treeview/utils'
